feat: 로그인

+++ src/main/java/sparta/auction_team_project/domain/user/repository/UserRepository.java
@@ -1,15 +1,10 @@
 package sparta.auction_team_project.domain.user.repository;
 
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
 import org.springframework.data.jpa.repository.JpaRepository;
 import sparta.auction_team_project.domain.user.entity.User;
-
-import java.math.BigInteger;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String username);
 
     boolean existsByEmail(String email);
